Changed constants pertaining to EMC to longs This helped find a few other places where it was incorrectly not using longs for EMC

+++ src/main/java/moze_intel/projecte/gameObjs/items/tools/PEToolBase.java
@@ -105,7 +105,7 @@ public abstract class PEToolBase extends ItemMode
 	/**
 	 * Clears the given OD name in an AOE. Charge affects the AOE. Optional per-block EMC cost.
 	 */
-	protected void clearOdAOE(World world, ItemStack stack, EntityPlayer player, String odName, int emcCost, EnumHand hand)
+	protected void clearOdAOE(World world, ItemStack stack, EntityPlayer player, String odName, long emcCost, EnumHand hand)
 	{
 		int charge = getCharge(stack);
 		if (charge == 0 || world.isRemote || ProjectEConfig.items.disableAllRadiusMining)
@@ -171,7 +171,7 @@ public abstract class PEToolBase extends ItemMode
 	/**
 	 * Tills in an AOE. Charge affects the AOE. Optional per-block EMC cost.
 	 */
-	protected void tillAOE(ItemStack stack, EntityPlayer player, World world, BlockPos pos, EnumFacing sidehit, int emcCost)
+	protected void tillAOE(ItemStack stack, EntityPlayer player, World world, BlockPos pos, EnumFacing sidehit, long emcCost)
 	{
 		int charge = this.getCharge(stack);
 		boolean hasAction = false;
@@ -309,7 +309,7 @@ public abstract class PEToolBase extends ItemMode
 	/**
 	 * Carves in an AOE. Charge affects the breadth and/or depth of the AOE. Optional per-block EMC cost.
 	 */
-	protected void digAOE(ItemStack stack, World world, EntityPlayer player, boolean affectDepth, int emcCost, EnumHand hand)
+	protected void digAOE(ItemStack stack, World world, EntityPlayer player, boolean affectDepth, long emcCost, EnumHand hand)
 	{
 		if (world.isRemote || this.getCharge(stack) == 0 || ProjectEConfig.items.disableAllRadiusMining)
 		{
@@ -381,7 +381,7 @@ public abstract class PEToolBase extends ItemMode
 	/**
 	 * Attacks in an AOE. Charge affects AOE, not damage (intentional). Optional per-entity EMC cost.
 	 */
-	protected void attackAOE(ItemStack stack, EntityPlayer player, boolean slayAll, float damage, int emcCost, EnumHand hand)
+	protected void attackAOE(ItemStack stack, EntityPlayer player, boolean slayAll, float damage, long emcCost, EnumHand hand)
 	{
 		if (player.getEntityWorld().isRemote)
 		{
@@ -442,7 +442,7 @@ public abstract class PEToolBase extends ItemMode
 	/**
 	 * Shears entities in an AOE. Charge affects AOE. Optional per-entity EMC cost.
 	 */
-	protected void shearEntityAOE(ItemStack stack, EntityPlayer player, int emcCost, EnumHand hand)
+	protected void shearEntityAOE(ItemStack stack, EntityPlayer player, long emcCost, EnumHand hand)
 	{
 		World world = player.getEntityWorld();
 		if (!world.isRemote)
